+++ src/core/Impl/startOf.ts
@@ -1,7 +1,7 @@
 import type { EsDay } from 'esday'
+import type { UnitWeek } from '~/common'
 import type { UnitType } from '~/types'
-import { prettyUnit } from '~/utils'
-import * as C from '../constant'
+import { C, prettyUnit } from '~/common'
 
 export function startOfImpl(that: EsDay, unit: UnitType, reverse = false) {
   const result = that.clone()
@@ -10,7 +10,7 @@ export function startOfImpl(that: EsDay, unit: UnitType, reverse = false) {
   // eslint-disable-next-line dot-notation
   const setterFunc = result['$set']
 
-  const instanceFactorySet = (method: 'h' | 'm' | 's' | 'ms', slice: number) => {
+  const instanceFactorySet = (method: Exclude<UnitType, UnitWeek>, slice: number) => {
     const argumentStart = [0, 0, 0, 0]
     const argumentEnd = [23, 59, 59, 999]
     const argument = reverse ? argumentEnd.slice(slice) : argumentStart.slice(slice)
@@ -18,7 +18,7 @@ export function startOfImpl(that: EsDay, unit: UnitType, reverse = false) {
   }
 
   const instanceFactory = (date: number, month: number) => {
-    setterFunc.call(result, 'M', [month, date])
+    setterFunc.call(result, C.MONTH, [month, date])
   }
 
   const $month = result.month()
@@ -28,11 +28,11 @@ export function startOfImpl(that: EsDay, unit: UnitType, reverse = false) {
   switch (prettyUnit(unit)) {
     case C.YEAR:
       reverse ? instanceFactory(31, 11) : instanceFactory(1, 0)
-      instanceFactorySet('h', 0)
+      instanceFactorySet(C.HOUR, 0)
       break
     case C.MONTH:
       reverse ? instanceFactory(0, $month + 1) : instanceFactory(1, $month)
-      instanceFactorySet('h', 0)
+      instanceFactorySet(C.HOUR, 0)
       break
     case C.WEEK:
     {
@@ -41,21 +41,21 @@ export function startOfImpl(that: EsDay, unit: UnitType, reverse = false) {
       const weekStart = C.INDEX_MONDAY
       const diff = ($day < weekStart ? $day + 7 : $day) - weekStart
       instanceFactory(reverse ? $date + (6 - diff) : $date - diff, $month)
-      instanceFactorySet('h', 0)
+      instanceFactorySet(C.HOUR, 0)
       break
     }
     case C.DAY:
     case C.DATE:
-      instanceFactorySet('h', 0)
+      instanceFactorySet(C.HOUR, 0)
       break
     case C.HOUR:
-      instanceFactorySet('m', 1)
+      instanceFactorySet(C.MIN, 1)
       break
     case C.MIN:
-      instanceFactorySet('s', 2)
+      instanceFactorySet(C.SECOND, 2)
       break
     case C.SECOND:
-      instanceFactorySet('ms', 3)
+      instanceFactorySet(C.MS, 3)
       break
   }
