+++ src/core/parseDate.ts
@@ -1,8 +1,7 @@
 import type { EsDay } from './EsDay'
 import type { DateType } from '~/types'
 import type { Tuple } from '~/types/util-types'
-import { isEmptyObject, isUndefined } from '~/utils'
-import * as C from './constant'
+import { C, isEmptyObject, isUndefined } from '~/common'
 
 export function parseArrayToDate(dateArray: number[]) {
   const dateArrayTuple: Tuple<number, 7> = [0, 0, 1, 0, 0, 0, 0]
