+++ src/types/index.ts
@@ -1,3 +1,3 @@
 export * from './core'
 
-export type { PrettyUnit, PrettyUnitType, UnitType } from '~/utils/unit'
+export type { PrettyUnit, PrettyUnitType, UnitType } from '~/common/unit'
